+++ app/prepends/prepended_controllers/with_stripped_down_file_set_id.rb
@@ -1,4 +1,4 @@
-module ImagesController
+module PrependedControllers::WithStrippedDownFileSetId
   def image_id
     params[:id].partition("/").first
   end
